[REFACTOR] Update GORM (#2)
* refactor account service

* refactore file service

* refactor notification service

* refactor seeder

* tidy dependencies

* add config seeder

+++ service/file/application/global/usecase_input.go
@@ -7,8 +7,8 @@ import (
 type FindAllInput struct {
 	QueryKey   string
 	QueryValue []interface{}
-	Limit      int64
-	Offset     int64
+	Limit      int
+	Offset     int
 	Sort       string
 }
